+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v0.5.0
+- 
+
 # v0.4.0
 - add config `source` functionality to update current sources
 - add config `source` functionality to new sources